optimisation/small-scalar-binding
All micro-benches give results we expect. Shout sum-check benches as
expected as well.
