+++ jolt-core/Cargo.toml
@@ -132,10 +132,6 @@ harness = false
 #[[bench]]
 #name = "shout_d_is_one_bench"
 #harness = false
-#
-#[[bench]]
-#name = "shout_d_greater_one_bench"
-#harness = false
 
 
 [[bench]]
