+++ jolt-core/src/field/mod.rs
@@ -1,10 +1,8 @@
 use std::fmt::{Debug, Display};
 use std::hash::Hash;
-use std::io::Read;
-use std::ops::{Add, AddAssign, Div, Mul, MulAssign, Neg, Sub, SubAssign};
 
-use allocative::Allocative;
-use ark_serialize::{CanonicalDeserialize, Valid, Validate};
+use std::ops::{Add, AddAssign, Div, Mul, MulAssign, Neg, Sub, SubAssign};
+use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
 use ark_std::{One, Zero};
 
 pub trait FieldOps<Rhs = Self, Output = Self>:
@@ -15,108 +13,7 @@ pub trait FieldOps<Rhs = Self, Output = Self>:
 {
 }
 
-/// A zero cost-wrapper around `u128` indicating that the value is already
-/// in Montgomery form for the target field
 
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, Default, Allocative)]
-pub struct MontU128(u128);
-
-impl From<u128> for MontU128 {
-    fn from(val: u128) -> Self {
-        // MontU128 can always be represented by 125 bits.
-        let val_masked = val & (u128::MAX >> 3);
-        MontU128(val_masked)
-    }
-}
-
-impl From<MontU128> for u128 {
-    fn from(val: MontU128) -> u128 {
-        val.0
-    }
-}
-impl From<u64> for MontU128 {
-    fn from(val: u64) -> Self {
-        MontU128::from(val as u128)
-    }
-}
-
-// === Serialization impls ===
-
-use ark_serialize::{CanonicalSerialize, Compress, SerializationError};
-use std::io::Write;
-
-impl CanonicalSerialize for MontU128 {
-    // fn serialize<W: Write>(&self, writer: W) -> Result<(), SerializationError> {
-    //     self.serialize_uncompressed(writer)
-    // }
-    fn serialize_with_mode<W: Write>(
-        &self,
-        writer: W,
-        _compress: Compress,
-    ) -> Result<(), SerializationError> {
-        self.serialize_uncompressed(writer)
-    }
-
-    fn serialized_size(&self, _compress: Compress) -> usize {
-        self.uncompressed_size()
-    }
-
-    fn serialize_compressed<W: Write>(&self, writer: W) -> Result<(), SerializationError> {
-        // u128 doesn’t really compress, so just write as 16 bytes
-        self.serialize_uncompressed(writer)
-    }
-
-    fn compressed_size(&self) -> usize {
-        16
-    }
-
-    fn serialize_uncompressed<W: Write>(&self, mut writer: W) -> Result<(), SerializationError> {
-        writer.write_all(&self.0.to_le_bytes())?;
-        Ok(())
-    }
-
-    fn uncompressed_size(&self) -> usize {
-        16
-    }
-}
-
-impl Valid for MontU128 {
-    fn check(&self) -> Result<(), SerializationError> {
-        Ok(())
-    }
-}
-
-impl CanonicalDeserialize for MontU128 {
-    fn deserialize_with_mode<R: Read>(
-        reader: R,
-        _compress: Compress,
-        _validate: Validate,
-    ) -> Result<Self, SerializationError> {
-        Self::deserialize_uncompressed(reader)
-    }
-
-    fn deserialize_compressed<R: Read>(reader: R) -> Result<Self, SerializationError> {
-        Self::deserialize_uncompressed(reader)
-    }
-
-    fn deserialize_compressed_unchecked<R: Read>(reader: R) -> Result<Self, SerializationError> {
-        Self::deserialize_uncompressed_unchecked(reader)
-    }
-
-    fn deserialize_uncompressed<R: Read>(mut reader: R) -> Result<Self, SerializationError> {
-        let mut buf = [0u8; 16];
-        reader.read_exact(&mut buf)?;
-        Ok(MontU128(u128::from_le_bytes(buf)))
-    }
-
-    fn deserialize_uncompressed_unchecked<R: Read>(
-        mut reader: R,
-    ) -> Result<Self, SerializationError> {
-        let mut buf = [0u8; 16];
-        reader.read_exact(&mut buf)?;
-        Ok(MontU128(u128::from_le_bytes(buf)))
-    }
-}
 
 pub trait JoltField:
     'static
@@ -313,5 +210,7 @@ where
 }
 
 pub mod ark;
+pub use ark::MontU128;
+
 // Commented out because of allocative issues for now
 // pub mod tracked_ark;
